feat(agent): observe what a turn actually spends
Rounds are the axis we bound and the one nobody reports a failure in. The
turn behind #8228 was four hours, 524 requests and 194M tokens; not one of
those numbers is a round count, and the same hundred rounds cost minutes or
hours depending on what each one read and how long the model thought.

This adds the axis that is missing: tokens, requests, wall clock and money,
accumulated per turn and reported after every round to sinks that opt into
event.RunBudgetSink. Cost goes through provider.Pricing, so a cache hit
prices at a fiftieth of a miss instead of reading as a full-price prompt —
on a cache-first agent, anything else would misreport the bill by an order
of magnitude.

Shadow only. Nothing reads a verdict from it and no threshold exists yet:
the distribution these samples record is what a threshold should be chosen
from, and picking one first would repeat how 100 was picked.

Rounds still count when a round's usage never arrived, so the axis never
reads cheaper than the turn was.

run_loop.go was 3 lines over the file-size ceiling with this, so the
sampling-usage accounting it already carried — estimate, merge, finalize,
emit — moved to run_usage.go as its own concern. That drops run_loop.go
under 800 lines and retires its file-size debt; the baseline diff carries
the two findings that moved with the extraction and records one fewer
finding overall.
