+++ tools/repolint/baseline.json
@@ -571,9 +571,11 @@
       "essay": 1
     },
     "internal/agent/run_loop.go": {
+      "essay": 27
+    },
+    "internal/agent/run_usage.go": {
       "complexity": 1,
-      "essay": 32,
-      "file-size": 179
+      "essay": 5
     },
     "internal/agent/save.go": {
       "complexity": 49,
